refactor: layout 디렉터리 Components 디렉토리 밖으로 뺌
