+++ pages/about.tsx
@@ -1,6 +1,6 @@
 import Link from "next/link";
 
-import RepresentativeLayout from "../components/layouts/RepresentativeLayout";
+import RepresentativeLayout from "../layouts/RepresentativeLayout";
 
 function AboutPage(): JSX.Element {
   return (
